requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
langchain
langchain-openai
langchain-pinecone
pinecone-client
python-dotenv